#96280 - [User] feature user profile

+++ internal/dtos/auth_dto.go
@@ -9,7 +9,7 @@ type LoginRequest struct {
 
 type LoginResponse struct {
 	User struct {
-		ID          int64  `json:"id"`
+		ID          uint   `json:"id"`
 		Name        string `json:"name"`
 		Email       string `json:"email"`
 		AccessToken string `json:"access_token"`
